Prepare for moving files automatically
