+++ packages/mui-system/src/breakpoints.d.ts
@@ -2,6 +2,7 @@ import { CSSObject } from '@mui/styled-engine';
 import { Breakpoints } from './createTheme/createBreakpoints';
 import type { Breakpoint } from './createTheme';
 import { ResponsiveStyleValue } from './styleFunctionSx';
+import { StyleFunction } from './Box';
 
 export interface ResolveBreakpointValuesOptions<T> {
   values: ResponsiveStyleValue<T>;
@@ -19,3 +20,12 @@ export function handleBreakpoints<Props>(
   propValue: any,
   styleFromPropValue: (value: any, breakpoint?: Breakpoint) => any,
 ): any;
+
+type DefaultBreakPoints = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+
+/**
+ * @returns An enhanced stylefunction that considers breakpoints
+ */
+export default function breakpoints<Props, Breakpoints extends string = DefaultBreakPoints>(
+  styleFunction: StyleFunction<Props>,
+): StyleFunction<Partial<Record<Breakpoints, Props>> & Props>;
