+++ packages/mui-system/src/compose.d.ts
@@ -1,0 +1,5 @@
+import { ComposedStyleFunction, StyleFunction } from './Box';
+
+export default function compose<T extends Array<StyleFunction<any>>>(
+  ...args: T
+): ComposedStyleFunction<T>;
