+++ packages/mui-system/src/cssGrid.d.ts
@@ -1,0 +1,3 @@
+import { PropsFor, grid } from './Box';
+
+export type CssGridProps = PropsFor<typeof grid>;
