+++ packages/mui-system/src/display.d.ts
@@ -1,0 +1,3 @@
+import { PropsFor, display } from './Box';
+
+export type DisplayProps = PropsFor<typeof display>;
