+++ packages/mui-system/src/flexbox.d.ts
@@ -1,0 +1,3 @@
+import { PropsFor, flexbox } from './Box';
+
+export type FlexboxProps = PropsFor<typeof flexbox>;
