+++ packages/mui-system/src/getThemeValue.d.ts
@@ -1,0 +1,1 @@
+export default function getThemeValue(prop: string, value: any, theme: object): any;
