+++ packages/mui-system/src/index.d.ts
@@ -1,91 +1,30 @@
-import {
-  ComposedStyleFunction,
-  StyleFunction,
-  PropsFor,
-  SimpleStyleFunction,
-  borders,
-  display,
-  flexbox,
-  grid,
-  palette,
-  positions,
-  shadows,
-  sizing,
-  typography,
-} from './Box';
 // disable automatic export
 export {};
 
-// borders.js
-export const border: SimpleStyleFunction<'border'>;
-export const borderTop: SimpleStyleFunction<'borderTop'>;
-export const borderRight: SimpleStyleFunction<'borderRight'>;
-export const borderBottom: SimpleStyleFunction<'borderBottom'>;
-export const borderLeft: SimpleStyleFunction<'borderLeft'>;
-export const borderColor: SimpleStyleFunction<'borderColor'>;
-export const borderTopColor: SimpleStyleFunction<'borderTopColor'>;
-export const borderRightColor: SimpleStyleFunction<'borderRightColor'>;
-export const borderBottomColor: SimpleStyleFunction<'borderBottomColor'>;
-export const borderLeftColor: SimpleStyleFunction<'borderLeftColor'>;
-export const borderRadius: SimpleStyleFunction<'borderRadius'>;
-export type BordersProps = PropsFor<typeof borders>;
+export * from './borders';
 
-// breakpoints.js
-type DefaultBreakPoints = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-export { handleBreakpoints, mergeBreakpointsInOrder } from './breakpoints';
+export { default as breakpoints, handleBreakpoints, mergeBreakpointsInOrder } from './breakpoints';
 
-/**
- * @returns An enhanced stylefunction that considers breakpoints
- */
-export function breakpoints<Props, Breakpoints extends string = DefaultBreakPoints>(
-  styleFunction: StyleFunction<Props>,
-): StyleFunction<Partial<Record<Breakpoints, Props>> & Props>;
+export { default as compose } from './compose';
 
-export function compose<T extends Array<StyleFunction<any>>>(...args: T): ComposedStyleFunction<T>;
+export * from './display';
 
-export type DisplayProps = PropsFor<typeof display>;
+export * from './flexbox';
 
-// flexbox.js
-export type FlexboxProps = PropsFor<typeof flexbox>;
+export * from './cssGrid';
 
-// grid.js
-export type GridProps = PropsFor<typeof grid>;
+export * from './palette';
 
-// palette.js
-export const color: SimpleStyleFunction<'color'>;
-export const bgcolor: SimpleStyleFunction<'bgcolor'>;
-export type PaletteProps = PropsFor<typeof palette>;
+export * from './positions';
 
-export type PositionsProps = PropsFor<typeof positions>;
+export * from './shadows';
 
-export type ShadowsProps = PropsFor<typeof shadows>;
+export * from './sizing';
 
-// * sizing.js TODO
-export const width: SimpleStyleFunction<'width'>;
-export const maxWidth: SimpleStyleFunction<'maxWidth'>;
-export const minWidth: SimpleStyleFunction<'minWidth'>;
-export const height: SimpleStyleFunction<'height'>;
-export const maxHeight: SimpleStyleFunction<'maxHeight'>;
-export const minHeight: SimpleStyleFunction<'minHeight'>;
-export const sizeWidth: SimpleStyleFunction<'sizeWidth'>;
-export const sizeHeight: SimpleStyleFunction<'sizeHeight'>;
-export const boxSizing: SimpleStyleFunction<'boxSizing'>;
-export type SizingProps = PropsFor<typeof sizing>;
+export * from './typography';
 
-// typography.js
-export const typographyVariant: SimpleStyleFunction<'typography'>;
-export const fontFamily: SimpleStyleFunction<'fontFamily'>;
-export const fontSize: SimpleStyleFunction<'fontSize'>;
-export const fontStyle: SimpleStyleFunction<'fontStyle'>;
-export const fontWeight: SimpleStyleFunction<'fontWeight'>;
-export const letterSpacing: SimpleStyleFunction<'letterSpacing'>;
-export const lineHeight: SimpleStyleFunction<'lineHeight'>;
-export const textAlign: SimpleStyleFunction<'textAlign'>;
-export const textTransform: SimpleStyleFunction<'textTransform'>;
-export type TypographyProps = PropsFor<typeof typography>;
+export { default as unstable_getThemeValue } from './getThemeValue';
 
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export function unstable_getThemeValue(prop: string, value: any, theme: object): any;
 /**
  * The `css` function accepts arrays as values for mobile-first responsive styles.
  * Note that this extends to non-theme values also. For example `display=['none', 'block']`
