+++ packages/mui-system/src/memoize.d.ts
@@ -1,0 +1,3 @@
+export default function memoize<ArgType = any, ReturnType = any>(
+  fn: (arg: ArgType) => ReturnType,
+): (arg: ArgType) => ReturnType;
