+++ packages/mui-system/src/positions.d.ts
@@ -1,0 +1,3 @@
+import { PropsFor, positions } from './Box';
+
+export type PositionsProps = PropsFor<typeof positions>;
