+++ packages/mui-system/src/shadows.d.ts
@@ -1,0 +1,3 @@
+import { PropsFor, shadows } from './Box';
+
+export type ShadowsProps = PropsFor<typeof shadows>;
