+++ packages/mui-system/src/sizing.d.ts
@@ -1,0 +1,12 @@
+import { PropsFor, SimpleStyleFunction, sizing } from './Box';
+
+export const width: SimpleStyleFunction<'width'>;
+export const maxWidth: SimpleStyleFunction<'maxWidth'>;
+export const minWidth: SimpleStyleFunction<'minWidth'>;
+export const height: SimpleStyleFunction<'height'>;
+export const maxHeight: SimpleStyleFunction<'maxHeight'>;
+export const minHeight: SimpleStyleFunction<'minHeight'>;
+export const sizeWidth: SimpleStyleFunction<'sizeWidth'>;
+export const sizeHeight: SimpleStyleFunction<'sizeHeight'>;
+export const boxSizing: SimpleStyleFunction<'boxSizing'>;
+export type SizingProps = PropsFor<typeof sizing>;
