+++ packages/mui-system/src/typography.d.ts
@@ -1,0 +1,12 @@
+import { PropsFor, SimpleStyleFunction, typography } from './Box';
+
+export const typographyVariant: SimpleStyleFunction<'typography'>;
+export const fontFamily: SimpleStyleFunction<'fontFamily'>;
+export const fontSize: SimpleStyleFunction<'fontSize'>;
+export const fontStyle: SimpleStyleFunction<'fontStyle'>;
+export const fontWeight: SimpleStyleFunction<'fontWeight'>;
+export const letterSpacing: SimpleStyleFunction<'letterSpacing'>;
+export const lineHeight: SimpleStyleFunction<'lineHeight'>;
+export const textAlign: SimpleStyleFunction<'textAlign'>;
+export const textTransform: SimpleStyleFunction<'textTransform'>;
+export type TypographyProps = PropsFor<typeof typography>;
